Convert the retool container into a StatefulSet
The Retool container is configured with a PersistentVolume. Adding a
PersistentVolume attachment to a deployment results in a deployment that
will fail to update. Converting it to a StatefulSet allows updates to
proceed without issue.

Alternatively, Retool probably *could* be run without a PersistentVolume
at all. The suggested use of the volume is to hold protobuf definitions
or other repositories of configuration data. Smaller configuration data
could be stored in a configmap; Most proto definitions will fit. Larger
configuration data could be mounted into a temporary volume via an
initcontainer.